+++ lib/assets/js/templates/navigation/js/index.js
@@ -392,7 +392,7 @@ System.register("node_modules/systemjs-plugin-babel/babel-helpers/createClass.js
     }
   };
 });
-System.register('templates/account/js/cookie.js', ['node_modules/systemjs-plugin-babel/babel-helpers/slicedToArray.js', 'node_modules/systemjs-plugin-babel/babel-helpers/classCallCheck.js', 'node_modules/systemjs-plugin-babel/babel-helpers/createClass.js'], function (_export, _context) {
+System.register('base/cookie.js', ['node_modules/systemjs-plugin-babel/babel-helpers/slicedToArray.js', 'node_modules/systemjs-plugin-babel/babel-helpers/classCallCheck.js', 'node_modules/systemjs-plugin-babel/babel-helpers/createClass.js'], function (_export, _context) {
     "use strict";
 
     var _slicedToArray, _classCallCheck, _createClass, Cookie;
@@ -416,7 +416,6 @@ System.register('templates/account/js/cookie.js', ['node_modules/systemjs-plugin
                     value: function getValue(key) {
                         var result = null;
 
-                        // TODO: re-write as this is inefficient
                         document.cookie.split('; ').forEach(function (cookie) {
                             var _cookie$split = cookie.split('=');
 
@@ -442,7 +441,7 @@ System.register('templates/account/js/cookie.js', ['node_modules/systemjs-plugin
         }
     };
 });
-System.register('templates/account/js/user.js', ['node_modules/systemjs-plugin-babel/babel-helpers/classCallCheck.js', 'node_modules/systemjs-plugin-babel/babel-helpers/createClass.js', '../../account/js/cookie'], function (_export, _context) {
+System.register('templates/account/js/user.js', ['node_modules/systemjs-plugin-babel/babel-helpers/classCallCheck.js', 'node_modules/systemjs-plugin-babel/babel-helpers/createClass.js', '../../../base/cookie'], function (_export, _context) {
     "use strict";
 
     var _classCallCheck, _createClass, cookie, MS_USER_COOKIE, User;
@@ -452,8 +451,8 @@ System.register('templates/account/js/user.js', ['node_modules/systemjs-plugin-b
             _classCallCheck = _node_modulesSystemjsPluginBabelBabelHelpersClassCallCheckJs.default;
         }, function (_node_modulesSystemjsPluginBabelBabelHelpersCreateClassJs) {
             _createClass = _node_modulesSystemjsPluginBabelBabelHelpersCreateClassJs.default;
-        }, function (_accountJsCookie) {
-            cookie = _accountJsCookie.default;
+        }, function (_baseCookie) {
+            cookie = _baseCookie.default;
         }],
         execute: function () {
             MS_USER_COOKIE = 'MS_USER_COOKIE';
