+++ lib/mns-core-ui/base/cookie.js
