+++ lib/mns-core-ui/templates/account/js/user.js
@@ -1,4 +1,4 @@
-import cookie from '../../account/js/cookie';
+import cookie from '../../../base/cookie';
 
 const MS_USER_COOKIE = 'MS_USER_COOKIE';
